+++ src/backend/joanie/core/models/products.py
@@ -39,8 +39,14 @@ from joanie.core.models.courses import (
     Enrollment,
     Organization,
 )
-from joanie.core.utils import contract_definition as contract_definition_utility
-from joanie.core.utils import issuers, webhooks
+from joanie.core.utils import (
+    contract_definition as contract_definition_utility,
+)
+from joanie.core.utils import (
+    issuers,
+    payment_schedule,
+    webhooks,
+)
 from joanie.core.utils.contract_definition import embed_images_in_context
 from joanie.core.utils.payment_schedule import generate as generate_payment_schedule
 from joanie.signature.backends import get_signature_backend
@@ -247,6 +253,23 @@ class Product(parler_models.TranslatableModel, BaseModel):
         """
         dates = self.get_equivalent_course_run_dates()
         return CourseRun.compute_state(**dates)
+
+    @property
+    def is_withdrawable(self):
+        """
+        Return True if the product has a withdrawal period.
+
+        Read the docstring of core.utils.payment_schedule.has_withdrawal_period method
+        for further information.
+        """
+        start_date = self.get_equivalent_course_run_dates().get("start")
+
+        if not start_date:
+            return True
+
+        return payment_schedule.has_withdrawal_period(
+            timezone.localdate(), start_date.date()
+        )
 
     def clean(self):
         """