+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - Add link to back office in Django admin topbar
 - Add boolean field `has_waived_withdrawal_right` to the `Order` model
+- Add `is_withdrawable` to `Product` model
 
 ## [2.9.2] - 2024-10-24